thesis: Deleted the font customization of nickie
I'm not quite sure what this customization tries to achieve but there
seems to be a problem with it; it uses a typewriter font as main font.
Thus,  I decided that, unless the \setromanfont feels inadequate, I will
erase nickie's customization (there are a whole lotta lines to comment
out).